+++ src/app/inobeta-ui/http/auth/session.stub.spec.ts
@@ -1,9 +1,6 @@
 import {of} from 'rxjs';
 import {fixtures} from './session.fixture.spec';
 import {authServiceStub} from './auth.service.stub.spec';
-import {localStorageStub} from './local-storage.stub.spec';
-import {cookiesStorageStub} from './cookies-storage.stub.spec';
-
 export const sessionServiceStub = {
 
   setAuthtype: () => {
@@ -16,8 +13,6 @@ export const sessionServiceStub = {
   },
 
   logout: () => {
-    localStorageStub.clear();
-    cookiesStorageStub.clear();
   }
 
 };