+++ src/app/inobeta-ui/http/store/loader/actions.ts
@@ -1,7 +1,7 @@
-import { createAction } from "@ngrx/store";
+import { createAction, props } from "@ngrx/store";
 
 export const ibLoaderActions = {
-    incLoading: createAction('[Http/Loader] incLoading'),
-    decLoading: createAction('[Http/Loader] decLoading'),
+    incLoading: createAction('[Http/Loader] incLoading', props<{ url: string, method: string}>()),
+    decLoading: createAction('[Http/Loader] decLoading', props<{ url: string, method: string}>()),
     skipShow: createAction('[Http/Loader] skipShow'),
 }