+++ src/app/inobeta-ui/http/store/loader/interfaces.ts
@@ -2,4 +2,5 @@ export interface IbLoaderState{
   pendingRequests: number;
   showLoading: boolean;
   skipShow: boolean;
+  pendingRequestList: { url: string, method: string}[]
 }