+++ src/app/inobeta-ui/http/store/loader/selectors.ts
@@ -11,3 +11,10 @@ export const ibSelectIsHttpLoading = createSelector(
     return state?.loader?.showLoading ?? false
   }
 )
+
+export const ibSelectIsHttpUrlLoading = (endpoint: { url: string, method: string}) => createSelector(
+  selectFeature,
+  (state: IHttpStore): boolean => {
+    return state?.loader?.pendingRequestList.findIndex(pl => pl.url === endpoint.url && pl.method === endpoint.method) >= 0 ?? false
+  }
+)